Refactoring and types extracting for Endpoint module
