+++ src/endpoint/index.ts
@@ -1,7 +1,5 @@
-import { AxiosResponse, AxiosError, AxiosInstance, AxiosRequestConfig } from 'axios';
-import { BindedActionPayload } from '@lib/redux-http';
-export type ResponseTransform = (input: AxiosResponse) => any;
-export type ErrorTransform = (input: AxiosError) => any;
+import { AxiosResponse, AxiosInstance, AxiosRequestConfig } from 'axios';
+import { ErrorTransform, ResponseTransform } from './requestHandler';
 
 export type requestHandler = (
   requestConfiguration: AxiosRequestConfig,
@@ -10,36 +8,5 @@ export type requestHandler = (
   axiosInstance: AxiosInstance,
 ) => Promise<AxiosResponse>;
 
-export type Method =
-  | 'get'
-  | 'delete'
-  | 'head'
-  | 'options'
-  | 'post'
-  | 'put'
-  | 'patch';
-
-export type EndpointConfig = {
-  path: string,
-  method: Method,
-  name: string,
-  errorTransformer?: ErrorTransform;
-  responseTransformer?: ResponseTransform,
-};
-
-export type BindRequest = {
-  config: EndpointConfig,
-  actionName: string,
-  axiosInstance: AxiosInstance,
-  signature: Symbol,
-};
-
-export type EndpointResponse = {
-  rawResponse: AxiosResponse | AxiosError,
-  data: any,
-  success: boolean,
-};
-
-export type BindedAction = (params: any) => BindedActionPayload;
-
-export { default as Endpoint } from '@lib/endpoint/Endpoint';
+export * from './Endpoint';
+export * from './requestHandler';
